requirements.txt
python-telegram-bot==21.6
flask==3.0.3
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
requests==2.32.3
python-dotenv==1.0.1
apscheduler==3.10.4
mercadopago==2.2.3
qrcode==7.4.2
pillow==10.4.0
pytz==2024.2
schedule==1.2.2
psutil==6.0.0
cryptography==43.0.1
gunicorn==21.2.0